Modified L1 spike delivery to better exploit shared memory staging buffers
Previously a single buffer was dedicated to a single global memory buffer slot.
This is wasteful if there are a small number of partitions and introduces a
sequential bottleneck. Now the available buffers are split between the total
number of partitions. Further savings would be possible for small number of
partitions by not dedicating buffers to self.

Also fixed a bug in allocation of buffer memory. Previously 4KB of data was
allocated but the code would easily use 4x this, which would introduce problems
for a large number of partitions.

The current code limits the number of partitions to 32.