Modified Axon data structure
Synapses are now ordered by source, delay, and then target. For each target (if
there are parallel synapses) synapses are stored LIFO. Previously all synapses
for a target was stored LIFO.

This change breaks some regression tests, where 'disconnect' has been used. The
reason for this is that the ordering of the synapses are different. If there
are parallel synapses the disconneted one may differ.

The file encode/decode magic numbers are changed due to the different data
structure, which breaks one regression.

A number of additional quickcheck tests have been added for Axon.hs