Changed format of reverse connectivity matrix
Previosly this was organised as the forward matrix, split first by neuron and
then by delay. For the reverse matrix we only ever access all the synapses of
the neuron at the same time, so storing it in this manner is wasteful. It is
now stored with all synapses in a single row.