Merged LTD and LTP into a single step
Both potentiation and depression is done at the same time at the very end of
the STDP window. This simplifies a fair amount of spike delivery code. There is
now just a single STDP accumulation buffer rather than separate ones for LTP
and LTD.  As a consequence of this there is now just one STDP application
kernel, which is also faster.