Build update: merged several binaries into single nemo binary
Previously we had these separate binaries: 'nemo', 'nemo-server', 'runtests',
and 'benchmark'. These are now all part of the 'nemo' binary, which can now be
called with a first paramter 'server', 'test', and 'benchmark'. The support for
these three extra execution mode is controlled via the same build flags as
before.

The examples were also separate programs before. These ought to be loadable
scripts, but for now they are folded into the main nemo program in the same
manner.

The build should now be faster, as the same files are not compiled several times.