Fixed build failure in emulation mode
The include chain for the kernel code ends up referencing shared_ptr which
makes use of exceptions. The compiler will not compile emulation kernel code
which relies on exceptions, and it's not able to tell that exceptions are not
actually used. Fixed this by chaning use of RSMatrix in ConnectivityMatrix to
pointers, thus breaking the include chain.