Made Matlab client connect directly to server
In previous versions the connection to the server was handled by an
intermediate nemo client. This added no functionality and only served to
consume memory. We may add this back later if we add functionality to the nemo
client itself.

Existing m-files may need to be modified:

- nemoSetBackend is no longer used; instead
- nemoConnect is used in the form nemoConnect(host, port); and
- nemoDisableSTDP is no longer used.

The frontend protocol specification is left in place.