CUDA: refactored L0 delivery
The previous version tried to deal with multiple delays for the same
presynaptic neuron at the same time. This would have some positive effect if

1. neurons fire frequently
2. neurons have synapses with different delays
3. each of those delays have few synapses

This made the L0 delivery more complex and seems to be the wrong thing for
which to optmise.

This commit replaces the above scheme with one where there's an extra loop
which iterates over all the relevant delay for a single firing presynaptic
neuron. In the common case there will be only one iteration of this loop.

For the benchmark program (1000 synapses with various delays) this speeds
things up by around 20% (for 2k, 4k, and 8k neurons).