CUDA: added basic warp-scheduling
The commit stage in the gather step is expensive due to the need to ensure
atomic update of the incoming current. We rely on warp-sized groups of synapses
not having any such conflicts, and then perform these updates in sequence. To
partially parallelise this, added a target warp indicator associated with each
group. This is a 32-bit vector indicating which warps are targeted by the
current group. Groups which have no conflict (determined using simple bit
operations), are scheduling to commit in parallel.

The current implementation does not optimise the ordering of synapses to
improve scheduling.