CUDA: reduced looping code in scatter step
Previously 'scatter' relied on the firing history being available. To detect
firing neurons it iterated through the firing history. Now, we set the firing
indices directly in the fire step, and use this in 'scatter'. Also moved
'scatter' to immideately after 'fire' which means we can later use the same 8KB
block of smem for firing history and current accumulation.

The effect on performance is not terribly large. The scatter step reduces from
492ms to 452ms for n=1024,m=1000. The main benefit is the opportunity for smem
savings, however.