CUDA: removed preprocessor insanity related to STDP
We previously generated two different kernels for STDP and non-STDP. Previous
version of the kernel had STDP-related code in many places and did different
processing depending on whether STDP was enabled or not. Now, all STDP-related
code is in one place, so we can simply use a conditional for this. The cost is
approximately 1M spike deliveries per second, which isn't much in the big
scheme of things.