CUDA: moved FCM host storage from split to unified format
Previously the FCM was stored in separate blocks for each synaspe group (source
partition/target partition/delay), as this was how the FCM was organised on the
device. We now have a single large FCM on the device, and to reflect this we
store the data the same way on the host side, thus simplifying the code.