CUDA: added generic connectivity matrix class
This makes no reference to cuda partitioning and is thus suitable for
construction of networks that will be later used in other backends. The new CM
class is used for construction, but the old code is left in place.

The CPU backend has been changed to use single-precision floating point, which
breaks CPU backend regression tests.