Added cuPrintf functionality
The cuPrintf source files are not included in the project since these are only
available to registered CUDA developers. The user can specify a directory in
which to find these files. Also tidied the logging code in the CUDA backend,
all of which is controlled with the cmake variable NEMO_CUDA_DEBUG_TRACE.