Cuda: modified firing buffer to sync with device every cycle
This is the common use-case so we should optimise for this rather than
long-running simulations that leave data on the device. The overhead compared
with reading only every 1000ms is minimal, e.g. throughput changes from 280M to
269M.

This commit also performs a small optimisation for the host-side processing of
the firing data. This means that reading back data every cycle is now faster
than reading back data every 1000ms in previous commit and with only a small
difference compared with not reading data back at all (one example: 384M vs
336M).