Cuda: modified low-level current stimulus API
The low-level setCurrentStimulus function now requires that the input current
vector has the exact same format as that used internally. This can be
accomplished by making use of the mapper that the backend can return. This is
how it is handled in the MPI worker and this eliminates a fair amount of
unneccessary busy-work mapping between index spaces. For 40k neurons and 4
workers, this accounts for 15% of execution time, and the removed portion of
the code scaled poorly.