Cuda: added multiple streams for processing and communication
This uses one stream for all kernels and one stream for the firing buffer copy.
Experiemented with a number of different approaches (one vs two streams,
different order of scatter and firing copy) and found this to be the fastest
and the only one that makes up for having split the kernel in the first place.

A few points to note:

- the stimulus copy is not part of this stream system
- some kernels (updateStdp and scatter) could be in different stream at the
  cost of more synchronisation.