Build: removed TIMING_ENABLED option
This was originally added in order to allow building the library without
requiring non-header boost libraries. At the time the 'date_time'
library was the only such boost library, and it is not strictly
necessary for NeMo. Now, however, we have a few other non-optional
non-header boost libraries, so this option is no longer useful.