Cuda: accumulated currents stored as floating point
Previously accumulated currents were computed and stored in gmem in
fixed-point format. Conversion to floating point was done in the neuron
plugin. Now, the conversion is done in gather, such that the neuron
plugin only needs to deal with floating point data.

Input currents are never converted to fixed point, which can cause very
slight differences when using external input currents. One of the
existing regression tests fails because of this.

The internal plugin API also changes in this commit